[refactor] Major project structure reorganization
