+++ src/dashboard/assets/img/sidebar_ellipse.png
